refactor: :truck: added ebook and abook folder for structure
